Allow downgrades to Airflow 2 without FAB provider (#55231)
We don't "always" need the FAB provider, we just need it if we don't
have the db at the version we expect. This can be because 1) we upgraded
from AF2 and the tables still exist or 2) the instance had the FAB
provider in the past but doesn't any longer. In either case though, we
have the tables, and at the version we expect anyway, so no need to
require the FAB provider to be installed before the downgrade.