Only use runner cleanup where it is really needed (#55345)
We decided to clean-up runner and remove all the potential garbage
we get when we start runner - we also move docker to a separate
/mnt partitiion available on the runners - because for unit testing
especially we need as much space as possible. Also when we build
images, it is best if we use clean, pristine docker engine located
on the /mnt folder.

All of our other jobs are ok with small disk space and sometimes that cleanup
can take up to a minute (usually it is 10-20s)

This change (experimentally) removes all cleanup from all jobs
except the unit tests and image building to see if we can get stability
and speed at the same time.