AIP-67 - Multi-team: Per team executor config (env var only) (#55003)
* Multi-team: per team executor config (env var only)

Configuration for teams can now be specified via environment variable
using the triple underscore syntax outlined in AIP-67. This applies to
any configuration, but specifically is required for executor based
configuration.

A small shim has been added to BaseExecutor to allow easier access to
team based config.

ECS executor is converted to this new shim as a proof of concept for the
mechanism.

* PR Feedback: comment fixup