Update executor loader cache (#55469)
The executor loader caches the values it reads from config in several
ways to make lookups very fast. Now with teams, we may have multiple
instances of the same executor so the caches must now be per-team.

Also disallow having more than one list of executors for each team.

Add tests to cover more of these edge cases and to harden test coverage
of the multi-team usecase in general.