Fix handling of `priority_weight` for Dag Processor callbacks (#55436)
This reverses the order in the DB query to fetch callbacks for Dag
Processor from ascending to descending. As far as I can tell, currently
the `priority_weight` for Dag Processor is always 10. See:
https://github.com/apache/airflow/blob/811d1e91c5a387a208eb3570b43363ecc137f4ba/airflow-core/src/airflow/callbacks/database_callback_sink.py#L38
So this is not going to have a change in user experience but if we
wanted to allow the user to specify it in the future, this would cause
confusion because it was prioritizing lower weights over higher weights,
which is the opposite of how it works for tasks.

* Change default priority_weight from 10 to 1