AIP-67 - Multi Team: Pass args/kwargs to super in CeleryExecutor (#56006)
* pass args/kwargs to super in Celery executors

This allows the team_name to be passed to the super class. This is low
hanging fruit to allow the Celery executor to be used for multi team testing.
More changes will be needed to allow the Celery executor to use team-based
config, but that will be done at a future time.

* Fix compat tests