Add English as a fallback locale in the UI (#56028)
I ran in to this locally _somehow_ (no one else was able to reproduce it) but
`i18n.language` was undefined which caused an exception in the LocalProvider
component, resulting in nothing showing.

This provides a useful fallback so we always have a language in case one isn't
defined or can't be determined.