Fix migration errors for Pydantic 2.12.0 (#56579)
Pydantic 2.12.0 implemented experimental Sentinel that requires
newer version of typing extensions. Our migration scripts however
downgrade airflow to 2.11.0 and since airflow 2.11 does not have
pydantic specified as required dependency, it does not downgrade
it - but it downgrades typing-extensions that are airflow dependency.

This causes a mismatch between expected version of typing extensions
by Pydantic (4.14.1) and the one that we have installed in airflow
2.11 (4.13.1). However - in fact - pydantic is a dependency of
Airflow 2.11 - becuase serialization uses pydantic serializer
in 2.11 and it fails being imported if typing extensions is too low.

This is only a problem when downgrading to Airflow 2.11 with constraints
when you do not specify pydantic as extra. This should be fixed in
2.11.1 as there constraints should include latest version of
typing-extension and pydantic.

For now - the fix is to add pydantic as extra when downgrading
airflow to 2.11.0